copia_de_la_copia/requirements.txt
flet>=0.1.58
Pillow>=10.0.0
